refactor: move test fixtures out of __mocks__ into src/test-utils
`__mocks__` is a directory name Jest reserves for manual mocks — a `__mocks__` folder
next to a module is what `jest.mock()` resolves to, and at the root it auto-mocks
node_modules packages. Nothing under `src/__mocks__` was a mock. All three files are
plain fixture data imported directly, and no `jest.mock()` call resolved to them, so the
name claimed a Jest behavior the code never used.

Pure file move plus import paths, no logic change.

- src/test-utils/{models,composites,factory}.ts: moved from src/__mocks__/models/. The
  files sit one directory shallower now, so their own relative imports drop a level.
- src/components/{Polystat,auto_font_scaler}.test.ts(x) and
  src/components/tooltips/Tooltip.test.tsx: import from the new location.

`src/test-utils` matches the convention used in grafana/grafana.
