+++ src/test-utils/composites.ts
@@ -1,5 +1,5 @@
 import { ShowTimestampFormats } from 'components/types';
-import { CompositeItemType } from '../../components/composites/types';
+import { CompositeItemType } from '../components/composites/types';
 
 export const compositeA: CompositeItemType = {
   name: 'composite-a',
